Create RpcEndpoint interface
